third_party/blink/web_tests/external/wpt/editing/run/undo-redo-after-mutation-expected.txt

This is a testharness.js-based test.
[PASS] Redo for execCommand("insertLineBreak") after removing prior sibling with DOM API after undoing
[PASS] Redo for execCommand("insertLineBreak") after appending new child with DOM API after undoing
[PASS] Redo for execCommand("insertLineBreak") after inserting new child with DOM API after undoing
[PASS] Redo for execCommand("insertLineBreak") after removing its container with DOM API after undoing
[PASS] Redo for execCommand("insertLineBreak") after removing <i> element following the container with DOM API after undoing
[FAIL] Redo for execCommand("insertLineBreak") between <b> and <i> after removing preceding <b> element with DOM API after undoing
  assert_equals: <br> element should be inserted between the <b> and <i> elements by execCommand("insertLineBreak") expected "<b>will be removed</b><br><i>abc</i>" but got "<b>will be removed<br></b><i>abc</i>"
[FAIL] Redo for execCommand("insertLineBreak") between <b> and <i> after after removing following <i> element with DOM API after undoing
  assert_equals: <br> element should be inserted between the <b> and <i> elements by execCommand("insertLineBreak") expected "<b>abc</b><br><i>will be removed</i>" but got "<b>abc<br></b><i>will be removed</i>"
Harness: the test ran to completion.
